requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
tensorflow==2.12.0
pillow==9.5.0
numpy==1.23.5
python-multipart==0.0.6
Jinja2==3.1.6
